Lots of changes:
Fixed the slider and animation that I accidentally broke. (main.js:399)

Now update the 3D charts the correct way, but updating the data and redrawing. (main.js:329-336).
-- Before, a new chart was being created over the old one. This would then brake on animation after a while.
-- This change increases the animation speed tremendously!

Updated the 3D plot redraw to use the transformations. (main.js:329-336)
-- Before, if the user rotated, panned, etc, the view would be lost on rerender. Now the view transformations are preserved.

Updated the time slider to reflect the current time as it animates (main.js:318)